refactor(base): change fields types of rotor settings struct

+++ pkg/base/models.go
@@ -2,6 +2,6 @@ package base
 
 type RotorSettings struct {
 	Name    string
-	Pos     int
-	RingPos int
+	Pos     byte
+	RingPos byte
 }
